app/backend/requirements.txt
fastapi==0.74.1
uvicorn==0.17.5
pydantic==1.9.0
tensorflow==2.8.0
pillow==9.0.1
